Get graph compontents + split into files
